Changed Guava Xbee to software serial and did some analog pin changes for battery and solar panel(not used)

+++ weatherbox_firmware/src/gen_guava/gg_dev_battery.h
@@ -13,5 +13,5 @@
 #ifndef GG_DEV_BATTERY_H
 #define GG_DEV_BATTERY_H
 void gg_dev_battery_open(void);
-int gg_dev_battery_read(void);
+uint16_t gg_dev_battery_read(void);
 #endif
